Deleting original emulator core and beginning again from scratch
My original approach turned out to be quite inefficient. Space Invaders
uses only a subset of the entire 8080 instruction set, but I was randomly
implementing the entire set of opcodes without considering whether the
ROM actually used them or not. Not only does this waste time, but I cannot
know whether those opcodes are implemented correctly since they are not used.

This commit deletes the original emulator core and starts again from scratch.
Initially, every opcode except NOP (0x00) will be reported as unimplemented,
and the disassembly of that opcode shown. Then I will implement that opcode
and repeat the procedure until all the opcodes used in the ROM are properly
implemented.